% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_dt_table.R
\name{ro_dt_table}
\alias{ro_dt_table}
\title{datatable table to be used with DT}
\usage{
ro_dt_table(data, caption, sorting = FALSE, pagelength = 10)
}
\arguments{
\item{data}{Data frame or tibble. The data to display in the table.}

\item{caption}{Character. Table caption/title (displayed above the table).}

\item{sorting}{Logical. Should column sorting be enabled? Default is \code{FALSE}.}

\item{pagelength}{Integer. Number of rows to display per page. Default is \code{10}.}
}
\value{
data table html
}
\description{
Rendering an accessible, keyboard-navigable (with customJS) data table
table with options for sorting and page length.
}
\examples{
# In your Shiny server:
if (interactive()) {
ro_dt_table(
  data = my_data,
  caption = "Aantal gevallen per groep",
  sorting = TRUE,
  pagelength = 20
)
}
}
\seealso{
Other DT:
\code{\link[=ro_dt_theme]{ro_dt_theme()}}
}
\concept{DT}
