% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_dt_theme.R
\name{ro_dt_theme}
\alias{ro_dt_theme}
\title{Custom DT Table CSS Styling}
\usage{
ro_dt_theme(base_family = "RijksoverheidSansWebText", sorting = FALSE)
}
\arguments{
\item{base_family}{Character. Font family to use for table text. Default is "Arial".}

\item{sorting}{Logical. If \code{TRUE}, column headers are left-aligned;
if \code{FALSE}, right-aligned. Default is \code{FALSE}.}
}
\value{
A \code{shiny.tags} object with a \code{<head>} tag containing
custom CSS for DT tables.
}
\description{
Returns a Shiny tags object containing custom CSS styles for DT tables,
following preliminary RIVM huisstijl.
}
\examples{
if (interactive()) {
  ui <- fluidPage(
    ro_dt_theme(base_family = "Arial", sorting = FALSE),
    DTOutput("mytable")
  )
  server <- function(input, output, session) {
    output$mytable <- renderDT(datatable(iris, options(list(ordering=FALSE))))
  }
  shinyApp(ui, server)
}

}
\seealso{
Other DT:
\code{\link[=ro_dt_table]{ro_dt_table()}}
}
\concept{DT}
