% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_gt_theme.R
\name{ro_gt_theme}
\alias{ro_gt_theme}
\title{Use a standardized theme for gt tables}
\usage{
ro_gt_theme(df, ROfont = FALSE, ...)
}
\arguments{
\item{df}{Dataframe.}

\item{ROfont}{Default is FALSE, will change to TRUE for next major release (1.0.0).
If Rijksoverheid font is not available on your device, Verdana will be used.}

\item{...}{For internal use. Leave empty.}
}
\value{
gt table object.
}
\description{
`r ro_group_badge('gt')`
This function implements a stylized theme for gt tables.
}
\examples{
\dontrun{
ro_gt_theme(df)
}
}
\concept{gt}
