% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_reactable_theme.R
\name{ro_reactable_css}
\alias{ro_reactable_css}
\title{Custom CSS for reactable tables}
\usage{
ro_reactable_css(sorting = TRUE)
}
\arguments{
\item{sorting}{Logical. If \code{TRUE}, custom sorting arrows are shown in the
column headers. If \code{FALSE}, only the default arrows are hidden.
Default \code{TRUE}.}
}
\value{
A \code{shiny.tag} with a \verb{<head>} tag that contains custom CSS
for reactable tables.
}
\description{
Returns a Shiny tags object with custom CSS for reactable tables,
with table styling according to the Rijksoverheid huisstijl.
}
\examples{
if (interactive()) {
  library(shiny)
  library(reactable)

  ui <- fluidPage(
    ro_reactable_css(sorting = TRUE),
    reactableOutput("mytable")
  )

  server <- function(input, output, session) {
    output$mytable <- renderReactable(
      reactable(
        iris,
        defaultColDef = colDef(align = "right", sortable = TRUE),
        theme = ro_reactable_theme(sorting = TRUE)
      )
    )
  }

  shinyApp(ui, server)
}
}
\seealso{
Other reactable:
\code{\link[=ro_reactable_theme]{ro_reactable_theme()}}
}
\concept{reactable}
