% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_reactable_theme.R
\name{ro_reactable_theme}
\alias{ro_reactable_theme}
\title{Reactable theme following Rijksoverheid huisstijl.}
\usage{
ro_reactable_theme(base_family = "RijksoverheidSansWebText", sorting = FALSE)
}
\arguments{
\item{base_family}{Character. Font family to use for table text. Default is
"RijksoverheidSansWebText". If not installed, Verdana is used, then Arial,
then the first font found on your device, whichever is installed first.}

\item{sorting}{Logical. If \code{TRUE}, extra space is added to the right
in the column header for sort arrows. Default \code{FALSE}.}
}
\value{
A \code{reactable::reactableTheme} object.
}
\description{
Creates a \code{reactableTheme} object with table styling according to the
Rijksoverheid huisstijl.
}
\seealso{
Other reactable:
\code{\link[=ro_reactable_css]{ro_reactable_css()}}
}
\concept{reactable}
