% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROvis.table-package.R
\docType{package}
\name{ROvis.table-package}
\alias{ROvis.table}
\alias{ROvis.table-package}
\title{ROvis.table: Accessible Data Tables with Rijksoverheid Styling}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides accessible, keyboard-navigable data tables and themed visualizations using the 'gt' and 'DT' packages, with Rijksoverheid styling guidelines.
}
\author{
\strong{Maintainer}: ROvis team \email{spin@rivm.nl}

Authors:
\itemize{
  \item ROvis team \email{spin@rivm.nl}
}

}
\keyword{internal}
